Moved mosaic plotting functions to visualisation utils data.py
